+++ snomed/com.b2international.snowowl.snomed.datastore/src/com/b2international/snowowl/snomed/datastore/request/rf2/SnomedRf2ExportRequest.java
@@ -36,6 +36,7 @@ import org.hibernate.validator.constraints.NotEmpty;
 
 import com.b2international.commons.CompareUtils;
 import com.b2international.commons.StringUtils;
+import com.b2international.commons.collections.Collections3;
 import com.b2international.snowowl.core.api.SnowowlRuntimeException;
 import com.b2international.snowowl.core.branch.Branch;
 import com.b2international.snowowl.core.date.DateFormats;
@@ -56,6 +57,7 @@ import com.b2international.snowowl.snomed.datastore.internal.rf2.SnomedExportCli
 import com.b2international.snowowl.snomed.datastore.internal.rf2.SnomedExportResult;
 import com.b2international.snowowl.snomed.datastore.internal.rf2.SnomedRf2ExportModel;
 import com.b2international.snowowl.snomed.datastore.request.SnomedRequests;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.common.collect.FluentIterable;
 import com.google.common.collect.ImmutableMap;
 
@@ -64,17 +66,39 @@ import com.google.common.collect.ImmutableMap;
  */
 final class SnomedRf2ExportRequest implements Request<BranchContext, UUID> {
 
+	private static final long serialVersionUID = 1L;
+	
+	@JsonProperty
 	@NotEmpty
 	private String codeSystem;
+	
+	@JsonProperty
 	private boolean includeUnpublished;
-	@NotNull
+	
+	@JsonProperty 
+	@NotNull 
 	private Rf2ReleaseType releaseType;
+	
+	@JsonProperty 
 	private boolean extensionOnly;
+	
+	@JsonProperty 
 	private String startEffectiveTime;
+	
+	@JsonProperty 
 	private String endEffectiveTime;
+	
+	@JsonProperty 
 	private Collection<String> modules;
+	
+	@JsonProperty
 	private String transientEffectiveTime;
+	
+	@JsonProperty
 	private String namespace;
+	
+	@JsonProperty
+	private Collection<String> refSets;
 
 	SnomedRf2ExportRequest() {}
 	
@@ -112,6 +136,10 @@ final class SnomedRf2ExportRequest implements Request<BranchContext, UUID> {
 
 	void setNamespace(String namespace) {
 		this.namespace = namespace;
+	}
+	
+	public void setRefSets(Collection<String> refSets) {
+		this.refSets = Collections3.toImmutableSet(refSets);
 	}
 	
 	@Override
@@ -191,6 +219,7 @@ final class SnomedRf2ExportRequest implements Request<BranchContext, UUID> {
 		
 		model.setCodeSystemShortName(codeSystem);
 		model.setExtensionOnly(extensionOnly);
+		model.getRefSetIds().addAll(refSets);
 
 		return model; 
 	}